Added requirements.txt for venv and Dockerfile for running the application in a (dev) container
To improve the development workflow a requirements.txt is added for installing all the dependencies in a virtual environment. Furthermore, a basic Dockerfile is included to be able to run the application in an isolated environment. In future, a docker compose file should be added with a mounted volume for use as a development container.