Refactored the EmployeeTable widget
The widget previously used a list of string to generate the table headers and magic number indexes of this list to update the table. This is approach is very brittle. The list of strings is replaced by a Tuple[str, Callable[[Employee], Any]] (i.e. a label and extractor function). This removes magic numbers and cleans up the table generation and update methods.