Refactoed the validation and repository logic
The UI, business logic, and persistance layers are now decoupled, with the application/services acting as the birdge between the two layers. Global UI contexts provide an interface to UI components, with a controller handling the link between the UI and the business logic. In future, these global contexts can be reduced to more local scope (see ContextProviders in React)